rename video-player-impl-web
